backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic-settings==2.5.2
dashscope==1.20.14
PyMuPDF==1.24.10
python-docx==1.1.2
python-multipart==0.0.9
websockets==13.0
